pyproject.toml
[tool.poetry]
name = "etl-scores-and-odds"
version = "0.1.0"
description = ""
authors = ["Keegan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jupyter = "^1.0.0"
pandas = "^2.2.2"
google-cloud-bigquery = "^3.25.0"
pydata-google-auth = "^1.8.2"
bs4 = "^0.0.2"
selenium = "^4.23.0"
certifi = "^2024.7.4"
pyarrow = "^17.0.0"
db-dtypes = "^1.2.0"
ipykernel = "^6.29.5"
lxml = "^5.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
